+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+✨ Renamed to Hugie 🐻
+
 ### Added
 - Add endpoint subcommand
 - Add endpoint test command
