+++ setup.cfg
@@ -1,5 +1,5 @@
 [metadata]
-name = hfie
+name = hugie
 version = 0.1.0
 author = Matthew Upson
 author_email = hi@mantisnlp.com
@@ -30,4 +30,4 @@ dev =
 
 [options.entry_points]
 console_scripts =
-    hfie = hfie.cli:app
+    hugie = hugie.cli:app
